Split API tests by provider
